Renamed RefreshMessagesSyncTask to InboxIdleSyncWorker. | #993
