+++ FlowCrypt/src/main/java/com/flowcrypt/email/jetpack/workmanager/sync/InboxIdleSyncWorker.kt
@@ -3,7 +3,7 @@
  * Contributors: DenBond7
  */
 
-package com.flowcrypt.email.api.email.sync.tasks
+package com.flowcrypt.email.jetpack.workmanager.sync
 
 import android.content.Context
 import androidx.work.Constraints
@@ -18,7 +18,6 @@ import com.flowcrypt.email.api.email.FoldersManager
 import com.flowcrypt.email.api.email.model.LocalFolder
 import com.flowcrypt.email.database.entity.AccountEntity
 import com.flowcrypt.email.database.entity.MessageEntity
-import com.flowcrypt.email.jetpack.workmanager.sync.BaseSyncWorker
 import com.flowcrypt.email.service.MessagesNotificationManager
 import com.flowcrypt.email.util.GeneralUtil
 import com.sun.mail.imap.IMAPFolder
@@ -40,7 +39,7 @@ import javax.mail.UIDFolder
  * Time: 17:12
  * E-mail: DenBond7@gmail.com
  */
-class RefreshMessagesSyncTask(context: Context, params: WorkerParameters) : BaseSyncWorker(context, params) {
+class InboxIdleSyncWorker(context: Context, params: WorkerParameters) : BaseSyncWorker(context, params) {
   private val notificationManager = MessagesNotificationManager(applicationContext)
 
   override suspend fun runIMAPAction(accountEntity: AccountEntity, store: Store) {
@@ -128,7 +127,7 @@ class RefreshMessagesSyncTask(context: Context, params: WorkerParameters) : Base
           .enqueueUniqueWork(
               GROUP_UNIQUE_TAG,
               ExistingWorkPolicy.REPLACE,
-              OneTimeWorkRequestBuilder<RefreshMessagesSyncTask>()
+              OneTimeWorkRequestBuilder<InboxIdleSyncWorker>()
                   .addTag(TAG_SYNC)
                   .setConstraints(constraints)
                   .build()
